back/requirements.txt
fastapi
uvicorn[standard]
mangum
pillow
torch
ultralytics
opencv-contrib-python-headless
scikit-image
numpy
python-multipart
